Refactored fetchinstance
