+++ src/app/api/instances/route.ts
@@ -2,87 +2,30 @@ import {
   EC2Client,
   DescribeInstancesCommand,
   DescribeInstancesCommandOutput,
-  Instance,
+  // Instance,
 } from "@aws-sdk/client-ec2";
 import { NextRequest, NextResponse } from "next/server";
 import { pollRabbitMQServerStatus } from "@/utils/RabbitMQ/serverStatus";
 import createInstance from "@/utils/AWS/EC2/createBrokerInstance";
+import { listInstances, formattedInstances } from "./service";
 import { getEC2Regions } from "@/utils/AWS/EC2/getEC2Regions";
 import eventEmitter from "@/utils/eventEmitter";
 import { deleteEvent } from "@/utils/eventBackups";
 
-// Define an interface that extends the AWS Instance with a non-optional region.
-export interface InstanceWithRegion extends Instance {
-  region: string;
-}
-
 export const GET = async () => {
-  const params = {
-    Filters: [
-      {
-        Name: "tag:Publisher",
-        Values: ["Rabbitory"],
-      },
-      {
-        Name: "instance-state-name",
-        Values: ["pending", "running", "stopping", "stopped", "shutting-down"],
-      },
-    ],
-  };
+  try {
+    const instances = await listInstances();
 
-  const regions = await getEC2Regions();
-  if (!regions || regions.length === 0) {
-    return new NextResponse("No regions found", { status: 404 });
+    const formatted = formattedInstances(instances);
+
+    return NextResponse.json(formatted);
+  } catch (error) {
+    console.error("Error fetching instances:", error);
+    return NextResponse.json(
+      { message: "Error fetching instances" },
+      { status: 500 }
+    );
   }
-
-  const command = new DescribeInstancesCommand(params);
-
-  const instancePromises: Promise<InstanceWithRegion[]>[] = regions.map(
-    async (region) => {
-      const ec2Client = new EC2Client({ region });
-      try {
-        const response = (await ec2Client.send(
-          command
-        )) as DescribeInstancesCommandOutput;
-
-        const regionInstances: InstanceWithRegion[] =
-          response.Reservations?.flatMap(
-            (reservation) =>
-              reservation.Instances?.map((instance) => ({
-                ...instance,
-                region: region!,
-              })) ?? []
-          ) ?? [];
-
-        return regionInstances;
-      } catch (error) {
-        console.error(`Error querying region ${region}:`, error);
-        return [];
-      }
-    }
-  );
-
-  const instances: InstanceWithRegion[] = (
-    await Promise.all(instancePromises)
-  ).flat();
-
-  const formattedInstances = instances
-    .map((instance) => {
-      if (!instance || !instance.Tags) {
-        console.error("Instance or tags not found");
-        return null;
-      }
-      const name = instance.Tags.find((tag) => tag.Key === "Name")?.Value || "";
-      return {
-        name,
-        id: instance.InstanceId,
-        region: instance.region,
-        state: instance.State?.Name,
-      };
-    })
-    .filter(Boolean);
-
-  return NextResponse.json(formattedInstances);
 };
 
 export const POST = async (request: NextRequest) => {
