+++ src/app/api/instances/types.ts
@@ -1,0 +1,5 @@
+import { Instance } from "@aws-sdk/client-ec2";
+
+export interface InstanceWithRegion extends Instance {
+  region: string;
+}
